+++ hw19-project-work/novel/novel-api/src/main/java/me/whsv26/novel/api/application/NovelServiceImpl.java
@@ -5,6 +5,7 @@ import me.whsv26.novel.api.domain.AuthorId;
 import me.whsv26.novel.api.domain.GenreId;
 import me.whsv26.novel.api.domain.Novel;
 import me.whsv26.novel.api.domain.NovelId;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -16,6 +17,8 @@ import java.util.List;
 public class NovelServiceImpl implements NovelService {
 
     private final NovelRepository novelRepository;
+
+    private final CurrentUserProvider currentUserProvider;
 
     private final Clock clock;
 
@@ -40,6 +43,7 @@ public class NovelServiceImpl implements NovelService {
     @Override
     public Novel update(NovelId id, String title, String synopsis, List<GenreId> genres, List<String> tags) {
         var novel = findNovelById(id);
+        ensureIsAuthorFor(currentUserProvider.getCurrentUser(), novel);
         novel.update(title, synopsis, genres, tags, clock);
         return novelRepository.save(novel);
     }
@@ -48,6 +52,7 @@ public class NovelServiceImpl implements NovelService {
     @Override
     public void delete(NovelId id) {
         var novel = findNovelById(id);
+        ensureIsAuthorFor(currentUserProvider.getCurrentUser(), novel);
         novel.delete(clock);
         novelRepository.delete(novel);
     }
@@ -55,5 +60,12 @@ public class NovelServiceImpl implements NovelService {
     private Novel findNovelById(NovelId id) {
         return novelRepository.findById(id)
             .orElseThrow(() -> new NovelNotFoundException(id));
+    }
+
+    private static void ensureIsAuthorFor(CurrentUser user, Novel novel) {
+        var isAuthor = novel.getAuthorId().value().equals(user.userId());
+        if (!isAuthor) {
+            throw new AccessDeniedException("Only author can modify the novel");
+        }
     }
 }
